Originally thought of using byte instead of int but realized int was the standard. Changed to match type being used!

+++ age-calculator/src/main/java/org/example/Main.java
@@ -8,7 +8,7 @@ public class Main {
         //prompts for the user to enter their age
         System.out.print("Enter your age: ");
         //storing as int instead of a double because not expecting decimals
-        int age = scanner.nextByte();
+        int age = scanner.nextInt();
 
         int difference = 100 - age;
         System.out.print("You are " + difference + " years away from being 100 years old!");
